requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==1.0.0
Django==4.1
django-widget-tweaks==1.4.12
gcloud==0.18.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
jws==0.1.3
oauth2client==4.1.3
Pillow==9.2.0
protobuf==4.21.5
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==2.0.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
whitenoise==6.2.0
gunicorn==20.1.0
